Fixed: Unit test compilation issue

+++ src/test/src/main.cpp
@@ -1,5 +1,3 @@
-#include "gaia/ecs/query.h"
-#include "gaia/util/logging.h"
 #define NOMINMAX
 
 #include <atomic>
@@ -201,8 +199,8 @@ TEST_CASE("StringLookupKey") {
 	char tmp1[MaxLen];
 	GAIA_STRFMT(tmp0, MaxLen, "%s", "some string");
 	GAIA_STRFMT(tmp1, MaxLen, "%s", "some string");
-	core::StringLookupKey<128> l0(tmp0, strlen(tmp0), 0);
-	core::StringLookupKey<128> l1(tmp1, strlen(tmp1), 0);
+	core::StringLookupKey<128> l0(tmp0, (uint32_t)strlen(tmp0), 0);
+	core::StringLookupKey<128> l1(tmp1, (uint32_t)strlen(tmp1), 0);
 	CHECK(l0.len() == l1.len());
 	// Two different addresses in memory have to return the same hash if the string is the same
 	CHECK(l0.hash() == l1.hash());
